+++ familyhelper-life-stack/src/main/java/com/dwarfeng/familyhelper/life/stack/bean/dto/PbFile.java
@@ -1,0 +1,51 @@
+package com.dwarfeng.familyhelper.life.stack.bean.dto;
+
+import com.dwarfeng.subgrade.stack.bean.dto.Dto;
+
+import java.util.Arrays;
+
+/**
+ * 个人最佳文件。
+ *
+ * @author DwArFeng
+ * @since 1.0.0
+ */
+public class PbFile implements Dto {
+
+    private static final long serialVersionUID = 618096534284705527L;
+
+    private String originName;
+    private byte[] content;
+
+    public PbFile() {
+    }
+
+    public PbFile(String originName, byte[] content) {
+        this.originName = originName;
+        this.content = content;
+    }
+
+    public String getOriginName() {
+        return originName;
+    }
+
+    public void setOriginName(String originName) {
+        this.originName = originName;
+    }
+
+    public byte[] getContent() {
+        return content;
+    }
+
+    public void setContent(byte[] content) {
+        this.content = content;
+    }
+
+    @Override
+    public String toString() {
+        return "PbFile{" +
+                "originName='" + originName + '\'' +
+                ", content=" + Arrays.toString(content) +
+                '}';
+    }
+}